+++ packages/idempotency/src/types/PersistenceLayer.ts
@@ -1,0 +1,7 @@
+type PersistenceLayerConfigureOptions = {
+  functionName?: string
+};
+
+export { 
+  PersistenceLayerConfigureOptions
+};